Move mempool tests into `tests::vector` sub-module
Make it consistent with other test modules and prepare for adding
property tests.
